[Refactor] Extension 분리
