+++ Dokgi/View/HomeView/ViewController/HomeViewController.swift
@@ -461,40 +461,40 @@ extension HomeViewController: UICollectionViewDelegateFlowLayout {
     }
 }
 
-extension UICollectionViewCell {
-    // 셀이 작아지게 설정
-    func transformToSmall() {
-        UIView.animate(withDuration: 0.2) {
-            self.transform = CGAffineTransform(scaleX: 0.9, y: 0.9)
-        }
-    }
-    
-    // 기본 셀 크기로 지정
-    func transformToStandard() {
-        UIView.animate(withDuration: 0.2) {
-            self.transform = CGAffineTransform.identity
-        }
-    }
-}
+//extension UICollectionViewCell {
+//    // 셀이 작아지게 설정
+//    func transformToSmall() {
+//        UIView.animate(withDuration: 0.2) {
+//            self.transform = CGAffineTransform(scaleX: 0.9, y: 0.9)
+//        }
+//    }
+//    
+//    // 기본 셀 크기로 지정
+//    func transformToStandard() {
+//        UIView.animate(withDuration: 0.2) {
+//            self.transform = CGAffineTransform.identity
+//        }
+//    }
+//}
 
-extension UIViewController {
-    func setFloatingButton() {
-        let floatButton = FloatButton()
-        view.addSubview(floatButton)
-        
-        floatButton.snp.makeConstraints {
-            $0.trailing.equalToSuperview().offset(-25)
-            $0.bottom.equalTo(self.view.safeAreaLayoutGuide).offset(-40)
-            $0.width.height.equalTo(70)
-        }
-        
-        floatButton.addTarget(self, action: #selector(didTapButton), for: .touchUpInside)
-    }
-    
-    @objc func didTapButton() {
-        let addVC = AddPassageViewController()
-        self.navigationController?.pushViewController(addVC, animated: true)
-    }
-}
+//extension UIViewController {
+//    func setFloatingButton() {
+//        let floatButton = FloatButton()
+//        view.addSubview(floatButton)
+//        
+//        floatButton.snp.makeConstraints {
+//            $0.trailing.equalToSuperview().offset(-25)
+//            $0.bottom.equalTo(self.view.safeAreaLayoutGuide).offset(-40)
+//            $0.width.height.equalTo(70)
+//        }
+//        
+//        floatButton.addTarget(self, action: #selector(didTapButton), for: .touchUpInside)
+//    }
+//    
+//    @objc func didTapButton() {
+//        let addVC = AddPassageViewController()
+//        self.navigationController?.pushViewController(addVC, animated: true)
+//    }
+//}
